requirements.txt
playwright==1.42.0
psycopg2-binary==2.9.9
pytest==8.0.0
requests==2.31.0
python-dotenv==1.0.1
